requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.8.0
scipy>=1.9.0
pandas>=2.0.0